requirements.txt
pytesseract
opencv-python
Pillow
numpy
openai
streamlit
scikit-learn
pandas
joblib
matplotlib
seaborn
lightgbm
xgboost
